+++ CHANGELOG.md
@@ -31,6 +31,8 @@
 - Guns enchanted with infinity now still need to be reloaded, but no longer consume bullets.
 - Guns enchanted with piercing now deal reduced damage with each pierce (-20% damage each pierce).
 - Increased pulse rifle cooldown (4 -> 10 ticks).
+- Increased pistol cooldown (4 -> 5 ticks).
+- Decreased pistol damage (5 -> 3 half-hearts).
 
 ## 🐞 Fixes
 - Fixed marine camps sometimes generating in water.